third_party/WebKit/LayoutTests/http/tests/workers/worker-importScripts-onerror-crossorigin.html

<!DOCTYPE html>
<html>
<head>
    <script>
        window.jsTestIsAsync = true;
        window.isOnErrorTest = true;
    </script>
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
    <script>
        description("Ensure that scripts imported into a Worker from cross-origin hosts trigger sanitized onerror messages.");

        var worker = new Worker('resources/worker-importscripts-onerror-crossorigin.js');

        var workerOnerror, pageOnerror;
        worker.onmessage = function (e) {
            workerOnerror = e.data;
            shouldBeEqualToString("workerOnerror.message", "Script error.");
            shouldBeEqualToString("workerOnerror.filename", "");
            shouldEvaluateTo("workerOnerror.lineno", 0);
            shouldEvaluateTo("workerOnerror.colno", 0);
            shouldBeNull("workerOnerror.error");
        };

        var onerrorMessage, onerrorURL, onerrorLine;
        worker.onerror = function (e) {
            pageOnerror = e;
            shouldBeEqualToString("pageOnerror.message", "Script error.");
            shouldBeEqualToString("pageOnerror.filename", "");
            shouldEvaluateTo("pageOnerror.lineno", 0);
            shouldEvaluateTo("pageOnerror.colno", 0);
            shouldBeNull("pageOnerror.error");

            e.preventDefault();
            finishJSTest();
        };
    </script>
</body>
</html>
